Add mock endpoint + delete and move mapper methods
